refactor: #2098 refactor sagas folder structure
