+++ packages/marketplace/src/components/pages/help/__tests__/help.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { shallow } from 'enzyme'
 import { Help } from '../help'
 
-jest.mock('../../../../core/router', () => ({
+jest.mock('@/core/router', () => ({
   history: {
     push: jest.fn(),
   },
