+++ packages/marketplace/src/components/ui/__tests__/categories-list.tsx
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme'
 
 import CategoriesList, { CategoriesListProps } from '@/components/ui/categories-list'
 import CategoryItem from '@/components/ui/category-item'
-import { appCategorieStub } from '../../../sagas/__stubs__/app-categories'
+import { appCategorieStub } from '@/sagas/__stubs__/app-categories'
 import { CategoryModel } from '@reapit/foundations-ts-definitions'
 
 const props: CategoriesListProps = {
