+++ packages/marketplace/src/components/ui/menu.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
 import { useLocation } from 'react-router'
 import { Menu as Sidebar, MenuConfig, ReapitLogo } from '@reapit/elements'
-import Routes from '../../constants/routes'
+import Routes from '@/constants/routes'
 import { Location } from 'history'
 import { FaCloud, FaCloudDownloadAlt, FaCog, FaClipboardList } from 'react-icons/fa'
 import { MdHelp } from 'react-icons/md'
