+++ packages/marketplace/src/reducers/__tests__/app-detail.ts
@@ -1,7 +1,7 @@
 import appDetailReducer, { defaultState } from '../app-detail'
-import { ActionType } from '../../types/core'
-import ActionTypes from '../../constants/action-types'
-import { appDetailDataStub } from '../../sagas/__stubs__/app-detail'
+import { ActionType } from '@/types/core'
+import ActionTypes from '@/constants/action-types'
+import { appDetailDataStub } from '@/sagas/__stubs__/app-detail'
 
 describe('app-detail reducer', () => {
   it('should return default state if action not matched', () => {
