+++ packages/marketplace/src/sagas/installations/app-installations.ts
@@ -1,8 +1,8 @@
 import { put, fork, all, call, takeLatest } from '@redux-saga/core/effects'
-import ActionTypes from '../constants/action-types'
-import { Action } from '../types/core'
-import { errorThrownServer } from '../actions/error'
-import errorMessages from '../constants/error-messages'
+import ActionTypes from '@/constants/action-types'
+import { Action } from '@/types/core'
+import { errorThrownServer } from '@/actions/error'
+import errorMessages from '@/constants/error-messages'
 import { appInstallationsSetFormState, UninstallParams, InstallParams } from '@/actions/installations'
 import { selectLoggedUserEmail } from '@/selector/client'
 import { logger } from '@reapit/utils'
