+++ packages/marketplace/src/sagas/web-component/index.ts
@@ -1,0 +1,1 @@
+export * from './web-component'
