+++ packages/marketplace/src/utils/__tests__/actions.ts
@@ -1,7 +1,7 @@
 import { actionCreator, isType } from '../actions'
-import ActionTypes from '../../constants/action-types'
+import ActionTypes from '@/constants/action-types'
 import { clientFetchAppSummary, clientFetchAppSummarySuccess } from '@/actions/apps'
-import { Action } from '../../types/core'
+import { Action } from '@/types/core'
 import { ClientAppSummaryParams } from '@/reducers/client/app-summary'
 
 describe('actions utils', () => {
