+++ packages/marketplace/src/utils/__tests__/launch-app.ts
@@ -1,7 +1,7 @@
 import { handleLaunchApp } from '../launch-app'
 import { AppSummaryModel } from '@reapit/foundations-ts-definitions'
 
-jest.mock('../../core/store')
+jest.mock('@/core/store')
 
 describe('handleLaunchApp', () => {
   it('should run correctly', () => {
